adjust supported selectors doc section for block formatting
This commit changes the supported selectors doc section to render in a
code block on pkgsite. Previously, the section was treated as a single
bullet point and mushed together.

One downside of the Universal Selector is that is considered a bullet in
GoDoc, which caused the formatting issue. This change also moves the
mentioning of it the introduction of the block to avoid the issue.